rag_system/requirements.txt
# RAG System Requirements
# University of Antwerp - Information Retrieval Assignment 3

# Core ML/NLP
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0

# OpenAI API
openai>=1.0.0

# Token counting
tiktoken>=0.5.0

# Data handling
pandas>=2.0.0

# Progress bars
tqdm>=4.65.0

# Testing
pytest>=7.4.0

# Environment variables
python-dotenv>=1.0.0

# Agentic orchestration (optional, for advanced workflow)
# Ref: https://pypi.org/project/langgraph/ (1.0.5 released Dec 13, 2025)
langgraph>=1.0.5
